Update: 투표 WebSocket 구독을 페이지 레벨로 수정
